[UR] Implement `getProviderNativeError` in CUDA adptr. tests (#18055)
This fixes debug builds of UR with CUDA enabled; in that configuration
the `test-adapter-cuda` emits a call to `umf::umf2urResult` which
requires an implementaiton of `umf::getProviderNativeError`.
